pyproject.toml
[project]
name = "end-t-end-ai-agent-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "sial-sanwal",email = "sialsanwal885@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0" 
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langgraph (>=0.2.73,<0.3.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "streamlit (>=1.42.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
